requirements.txt
argcomplete==2.0.0
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
decorator==5.1.1
Django==4.0.3
django-appconf==1.0.5
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-imagekit==4.1.0
django-js-asset==2.0.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dynaconf==3.1.7
gunicorn==20.1.0
hvac==0.11.2
idna==3.3
knox==0.1.14
loguru==0.6.0
mysqlclient==2.1.0
pilkit==2.0
Pillow==9.0.1
pipx==1.0.0
protobuf==3.19.4
pycparser==2.21
PyJWT==2.3.0
pytz==2022.1
requests==2.27.1
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
userpath==1.8.0
validators==0.18.2
whitenoise==6.0.0
